rename FNNLayer to Forward Layer and put it in separate File
