+++ src/safeds/ml/nn/__init__.py
@@ -1,10 +1,10 @@
 """Classes for classification tasks."""
 
-from ._fnn_layer import FNNLayer
+from ._forward_layer import ForwardLayer
 from ._model import NeuralNetworkClassifier, NeuralNetworkRegressor
 
 __all__ = [
-    "FNNLayer",
+    "ForwardLayer",
     "NeuralNetworkClassifier",
     "NeuralNetworkRegressor",
 ]
