+++ tests/safeds/ml/nn/test_forward_layer.py
@@ -1,6 +1,6 @@
 import pytest
 from safeds.exceptions import OutOfBoundsError
-from safeds.ml.nn import FNNLayer
+from safeds.ml.nn import ForwardLayer
 
 
 @pytest.mark.parametrize(
@@ -15,7 +15,7 @@ def test_should_raise_if_input_size_out_of_bounds(input_size: int) -> None:
         OutOfBoundsError,
         match=rf"input_size \(={input_size}\) is not inside \[1, \u221e\)\.",
     ):
-        FNNLayer(output_size=1, input_size=input_size)
+        ForwardLayer(output_size=1, input_size=input_size)
 
 
 @pytest.mark.parametrize(
@@ -27,7 +27,7 @@ def test_should_raise_if_input_size_out_of_bounds(input_size: int) -> None:
     ids=["one", "twenty"],
 )
 def test_should_raise_if_input_size_doesnt_match(input_size: int) -> None:
-    assert FNNLayer(output_size=1, input_size=input_size).input_size == input_size
+    assert ForwardLayer(output_size=1, input_size=input_size).input_size == input_size
 
 
 @pytest.mark.parametrize(
@@ -42,7 +42,7 @@ def test_should_raise_if_unknown_activation_function_is_passed(activation_functi
         ValueError,
         match=rf"Unknown Activation Function: {activation_function}",
     ):
-        FNNLayer(output_size=1, input_size=1)._get_internal_layer(activation_function)
+        ForwardLayer(output_size=1, input_size=1)._get_internal_layer(activation_function)
 
 
 @pytest.mark.parametrize(
@@ -57,7 +57,7 @@ def test_should_raise_if_output_size_out_of_bounds(output_size: int) -> None:
         OutOfBoundsError,
         match=rf"output_size \(={output_size}\) is not inside \[1, \u221e\)\.",
     ):
-        FNNLayer(output_size=output_size, input_size=1)
+        ForwardLayer(output_size=output_size, input_size=1)
 
 
 @pytest.mark.parametrize(
@@ -69,4 +69,4 @@ def test_should_raise_if_output_size_out_of_bounds(output_size: int) -> None:
     ids=["one", "twenty"],
 )
 def test_should_raise_if_output_size_doesnt_match(output_size: int) -> None:
-    assert FNNLayer(output_size=output_size, input_size=1).output_size == output_size
+    assert ForwardLayer(output_size=output_size, input_size=1).output_size == output_size
