+++ tests/safeds/ml/nn/test_model.py
@@ -1,7 +1,7 @@
 import pytest
 from safeds.data.tabular.containers import Table, TaggedTable
 from safeds.exceptions import ModelNotFittedError, OutOfBoundsError, TestTrainDataMismatchError, InputSizeError
-from safeds.ml.nn import FNNLayer, NeuralNetworkClassifier, NeuralNetworkRegressor
+from safeds.ml.nn import ForwardLayer, NeuralNetworkClassifier, NeuralNetworkRegressor
 
 
 class TestClassificationModel:
@@ -17,7 +17,7 @@ class TestClassificationModel:
             OutOfBoundsError,
             match=rf"epoch_size \(={epoch_size}\) is not inside \[1, \u221e\)\.",
         ):
-            NeuralNetworkClassifier([FNNLayer(1, 1)]).fit(
+            NeuralNetworkClassifier([ForwardLayer(1, 1)]).fit(
                 Table.from_dict({"a": [1], "b": [2]}).tag_columns("a"),
                 epoch_size=epoch_size,
             )
@@ -34,14 +34,14 @@ class TestClassificationModel:
             OutOfBoundsError,
             match=rf"batch_size \(={batch_size}\) is not inside \[1, \u221e\)\.",
         ):
-            NeuralNetworkClassifier([FNNLayer(input_size=1, output_size=1)]).fit(
+            NeuralNetworkClassifier([ForwardLayer(input_size=1, output_size=1)]).fit(
                 Table.from_dict({"a": [1], "b": [2]}).tag_columns("a"),
                 batch_size=batch_size,
             )
 
     def test_should_raise_if_fit_function_returns_wrong_datatype(self) -> None:
         fitted_model = NeuralNetworkClassifier(
-            [FNNLayer(input_size=1, output_size=8), FNNLayer(output_size=1)],
+            [ForwardLayer(input_size=1, output_size=8), ForwardLayer(output_size=1)],
         ).fit(
             Table.from_dict({"a": [1], "b": [0]}).tag_columns("a"),
         )
@@ -49,7 +49,7 @@ class TestClassificationModel:
 
     def test_should_raise_if_predict_function_returns_wrong_datatype(self) -> None:
         fitted_model = NeuralNetworkClassifier(
-            [FNNLayer(input_size=1, output_size=8), FNNLayer(output_size=1)],
+            [ForwardLayer(input_size=1, output_size=8), ForwardLayer(output_size=1)],
         ).fit(
             Table.from_dict({"a": [1, 0], "b": [0, 1]}).tag_columns("a"),
         )
@@ -58,7 +58,7 @@ class TestClassificationModel:
 
     def test_should_raise_if_predict_function_returns_wrong_datatype_for_multiclass_classification(self) -> None:
         fitted_model = NeuralNetworkClassifier(
-            [FNNLayer(input_size=1, output_size=8), FNNLayer(output_size=3)],
+            [ForwardLayer(input_size=1, output_size=8), ForwardLayer(output_size=3)],
         ).fit(
             Table.from_dict({"a": [0, 1, 2], "b": [0, 15, 51]}).tag_columns("a"),
         )
@@ -67,12 +67,12 @@ class TestClassificationModel:
 
     def test_should_raise_if_model_has_not_been_fitted(self) -> None:
         with pytest.raises(ModelNotFittedError, match="The model has not been fitted yet."):
-            NeuralNetworkClassifier([FNNLayer(input_size=1, output_size=1)]).predict(
+            NeuralNetworkClassifier([ForwardLayer(input_size=1, output_size=1)]).predict(
                 Table.from_dict({"a": [1]}),
             )
 
     def test_should_raise_if_is_fitted_is_set_correctly_for_binary_classification(self) -> None:
-        model = NeuralNetworkClassifier([FNNLayer(input_size=1, output_size=1)])
+        model = NeuralNetworkClassifier([ForwardLayer(input_size=1, output_size=1)])
         assert not model.is_fitted
         model = model.fit(
             Table.from_dict({"a": [1], "b": [0]}).tag_columns("a"),
@@ -80,7 +80,7 @@ class TestClassificationModel:
         assert model.is_fitted
 
     def test_should_raise_if_is_fitted_is_set_correctly_for_multiclass_classification(self) -> None:
-        model = NeuralNetworkClassifier([FNNLayer(input_size=1, output_size=1), FNNLayer(output_size=3)])
+        model = NeuralNetworkClassifier([ForwardLayer(input_size=1, output_size=1), ForwardLayer(output_size=3)])
         assert not model.is_fitted
         model = model.fit(
             Table.from_dict({"a": [1, 0, 2], "b": [0, 15, 5]}).tag_columns("a"),
@@ -88,7 +88,7 @@ class TestClassificationModel:
         assert model.is_fitted
 
     def test_should_raise_if_test_and_train_data_mismatch(self) -> None:
-        model = NeuralNetworkClassifier([FNNLayer(input_size=1, output_size=1), FNNLayer(output_size=3)])
+        model = NeuralNetworkClassifier([ForwardLayer(input_size=1, output_size=1), ForwardLayer(output_size=3)])
         model = model.fit(
             Table.from_dict({"a": [1, 0, 2], "b": [0, 15, 5]}).tag_columns("a"),
         )
@@ -101,7 +101,7 @@ class TestClassificationModel:
             )
 
     def test_should_raise_if_table_size_and_input_size_mismatch(self) -> None:
-        model = NeuralNetworkClassifier([FNNLayer(input_size=1, output_size=1), FNNLayer(output_size=3)])
+        model = NeuralNetworkClassifier([ForwardLayer(input_size=1, output_size=1), ForwardLayer(output_size=3)])
         with pytest.raises(
             InputSizeError,
         ):
@@ -110,7 +110,7 @@ class TestClassificationModel:
             )
 
     def test_should_raise_if_fit_doesnt_batch_callback(self) -> None:
-        model = NeuralNetworkClassifier([FNNLayer(input_size=1, output_size=1)])
+        model = NeuralNetworkClassifier([ForwardLayer(input_size=1, output_size=1)])
 
         class Test:
             self.was_called = False
@@ -128,7 +128,7 @@ class TestClassificationModel:
         assert obj.callback_was_called() is True
 
     def test_should_raise_if_fit_doesnt_epoch_callback(self) -> None:
-        model = NeuralNetworkClassifier([FNNLayer(input_size=1, output_size=1)])
+        model = NeuralNetworkClassifier([ForwardLayer(input_size=1, output_size=1)])
 
         class Test:
             self.was_called = False
@@ -159,7 +159,7 @@ class TestRegressionModel:
             OutOfBoundsError,
             match=rf"epoch_size \(={epoch_size}\) is not inside \[1, \u221e\)\.",
         ):
-            NeuralNetworkRegressor([FNNLayer(input_size=1, output_size=1)]).fit(
+            NeuralNetworkRegressor([ForwardLayer(input_size=1, output_size=1)]).fit(
                 Table.from_dict({"a": [1], "b": [2]}).tag_columns("a"),
                 epoch_size=epoch_size,
             )
@@ -176,19 +176,19 @@ class TestRegressionModel:
             OutOfBoundsError,
             match=rf"batch_size \(={batch_size}\) is not inside \[1, \u221e\)\.",
         ):
-            NeuralNetworkRegressor([FNNLayer(input_size=1, output_size=1)]).fit(
+            NeuralNetworkRegressor([ForwardLayer(input_size=1, output_size=1)]).fit(
                 Table.from_dict({"a": [1], "b": [2]}).tag_columns("a"),
                 batch_size=batch_size,
             )
 
     def test_should_raise_if_fit_function_returns_wrong_datatype(self) -> None:
-        fitted_model = NeuralNetworkRegressor([FNNLayer(input_size=1, output_size=1)]).fit(
+        fitted_model = NeuralNetworkRegressor([ForwardLayer(input_size=1, output_size=1)]).fit(
             Table.from_dict({"a": [1], "b": [2]}).tag_columns("a"),
         )
         assert isinstance(fitted_model, NeuralNetworkRegressor)
 
     def test_should_raise_if_predict_function_returns_wrong_datatype(self) -> None:
-        fitted_model = NeuralNetworkRegressor([FNNLayer(input_size=1, output_size=1)]).fit(
+        fitted_model = NeuralNetworkRegressor([ForwardLayer(input_size=1, output_size=1)]).fit(
             Table.from_dict({"a": [1], "b": [2]}).tag_columns("a"),
         )
         predictions = fitted_model.predict(Table.from_dict({"b": [1]}))
@@ -196,12 +196,12 @@ class TestRegressionModel:
 
     def test_should_raise_if_model_has_not_been_fitted(self) -> None:
         with pytest.raises(ModelNotFittedError, match="The model has not been fitted yet."):
-            NeuralNetworkRegressor([FNNLayer(input_size=1, output_size=1)]).predict(
+            NeuralNetworkRegressor([ForwardLayer(input_size=1, output_size=1)]).predict(
                 Table.from_dict({"a": [1]}),
             )
 
     def test_should_raise_if_is_fitted_is_set_correctly(self) -> None:
-        model = NeuralNetworkRegressor([FNNLayer(input_size=1, output_size=1)])
+        model = NeuralNetworkRegressor([ForwardLayer(input_size=1, output_size=1)])
         assert not model.is_fitted
         model = model.fit(
             Table.from_dict({"a": [1], "b": [0]}).tag_columns("a"),
@@ -209,7 +209,7 @@ class TestRegressionModel:
         assert model.is_fitted
 
     def test_should_raise_if__test_and_train_data_mismatch(self) -> None:
-        model = NeuralNetworkRegressor([FNNLayer(input_size=1, output_size=1)])
+        model = NeuralNetworkRegressor([ForwardLayer(input_size=1, output_size=1)])
         model = model.fit(
             Table.from_dict({"a": [1, 0, 2], "b": [0, 15, 5]}).tag_columns("a"),
         )
@@ -222,7 +222,7 @@ class TestRegressionModel:
             )
 
     def test_should_raise_if_table_size_and_input_size_mismatch(self) -> None:
-        model = NeuralNetworkRegressor([FNNLayer(input_size=1, output_size=1), FNNLayer(output_size=3)])
+        model = NeuralNetworkRegressor([ForwardLayer(input_size=1, output_size=1), ForwardLayer(output_size=3)])
         with pytest.raises(
             InputSizeError,
         ):
@@ -231,7 +231,7 @@ class TestRegressionModel:
             )
 
     def test_should_raise_if_fit_doesnt_batch_callback(self) -> None:
-        model = NeuralNetworkRegressor([FNNLayer(input_size=1, output_size=1)])
+        model = NeuralNetworkRegressor([ForwardLayer(input_size=1, output_size=1)])
 
         class Test:
             self.was_called = False
@@ -249,7 +249,7 @@ class TestRegressionModel:
         assert obj.callback_was_called() is True
 
     def test_should_raise_if_fit_doesnt_epoch_callback(self) -> None:
-        model = NeuralNetworkRegressor([FNNLayer(input_size=1, output_size=1)])
+        model = NeuralNetworkRegressor([ForwardLayer(input_size=1, output_size=1)])
 
         class Test:
             self.was_called = False
